[Gluten-1205][VL] Merge GlutenColumnarBatchSerializer and CelebornColumnarBatchSerializer
